Исправление замечаний по ревью

+++ src/main/java/ru/yandex/practicum/filmorate/service/impl/FriendshipServiceImpl.java
@@ -5,6 +5,7 @@ import org.springframework.stereotype.Service;
 import ru.yandex.practicum.filmorate.entity.User;
 import ru.yandex.practicum.filmorate.service.interfaces.FriendshipService;
 import ru.yandex.practicum.filmorate.storage.friendship.FriendshipStorage;
+import ru.yandex.practicum.filmorate.storage.user.UserStorage;
 
 import java.util.List;
 
@@ -13,15 +14,20 @@ import java.util.List;
 public class FriendshipServiceImpl implements FriendshipService {
 
     private final FriendshipStorage friendshipStorage;
+    private final UserStorage userStorage;
 
     @Override
     public void addNewFriend(long userId, long friendId) {
-        friendshipStorage.addNewFriend(userId, friendId);
+        if (userStorage.getUserById(userId) != null && userStorage.getUserById(friendId) != null) {
+            friendshipStorage.addNewFriend(userId, friendId);
+        }
     }
 
     @Override
     public void deleteFromFriends(long userId, long friendId) {
-        friendshipStorage.deleteFromFriends(userId, friendId);
+        if (userStorage.getUserById(userId) != null && userStorage.getUserById(friendId) != null) {
+            friendshipStorage.deleteFromFriends(userId, friendId);
+        }
     }
 
     @Override
